requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
python-dotenv
flake8
psycopg2
pandas
httpx
firebase-admin
pytz

### RabbitMQ ###
paho-mqtt

### PostgreSQL ###
SQLAlchemy

### MongoDB ###
pymongo
